+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildTarMojo.java
@@ -89,8 +89,8 @@ public class BuildTarMojo extends JibPluginConfiguration {
               getProject().getName(),
               getProject().getVersion(),
               mavenHelpfulSuggestionsBuilder.build());
-      Path tarOutputPath =
-          Paths.get(getProject().getBuild().getDirectory()).resolve("jib-image.tar");
+      Path buildOutput = Paths.get(getProject().getBuild().getDirectory());
+      Path tarOutputPath = buildOutput.resolve("jib-image.tar");
       TarImage targetImage = TarImage.named(targetImageReference).saveTo(tarOutputPath);
 
       PluginConfigurationProcessor pluginConfigurationProcessor =
@@ -112,6 +112,7 @@ public class BuildTarMojo extends JibPluginConfiguration {
               .build();
 
       BuildStepsRunner.forBuildTar(tarOutputPath)
+          .writeImageDigest(buildOutput.resolve("jib-image.digest"))
           .build(
               jibContainerBuilder,
               containerizer,